Routes and bugs

+++ app/Http/Controllers/Admin/BookController.php
@@ -3,15 +3,14 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
-use App\Http\Requests\BookRequest;
+use App\Http\Requests\UpdateBookRequest;
+use App\Http\Requests\StoreBookRequest;
 use App\Http\Resources\BookResource;
 use App\Models\Book;
 use Illuminate\Http\Request;
+
 class BookController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     */
     public function index(Request $request)
     {
         $query = Book::query();
@@ -24,41 +23,55 @@ class BookController extends Controller
 
         $books = $query->paginate(10);
 
+        if ($books->isEmpty()) {
+            return response()->json([
+                'message' => 'هیچ کتابی یافت نشد.'
+            ], 404);
+        }
+
         return BookResource::collection($books);
     }
 
-    /**
-     * Store a newly created resource in storage.
-     */
-    public function store(BookRequest $request)
+    public function store(StoreBookRequest $request)
     {
         $book = Book::create($request->validated());
+        return (new BookResource($book))
+            ->response()
+            ->setStatusCode(201);
+    }
+
+    public function show($id)
+    {
+        $book = Book::find($id);
+
+        if (! $book) {
+            return response()->json([
+                'message' => 'کتاب مورد نظر یافت نشد.'
+            ], 404);
+        }
+
         return new BookResource($book);
     }
 
-    /**
-     * Display the specified resource.
-     */
-    public function show(Book $book)
-    {
-        return new BookResource($book);
-    }
-
-    /**
-     * Update the specified resource in storage.
-     */
-    public function update(BookRequest $request, Book $book)
-    {
+    public function update(UpdateBookRequest $request, Book $book) {
         $book->update($request->validated());
         return new BookResource($book);
     }
 
-    /**
-     * Remove the specified resource from storage.
-     */
-    public function destroy(Book $book)
+    public function destroy($id)
     {
+        $book = Book::find($id);
+
+        if (! $book) {
+            return response()->json([
+                'message' => 'کتاب مورد نظر برای حذف یافت نشد.'
+            ], 404);
+        }
+
         $book->delete();
-        return response()->json(['message'=>'book deleted successfully']);
+
+        return response()->json([
+            'message' => 'کتاب با موفقیت حذف شد.'
+        ], 200);
     }
 }
